Move ReadDir to new iters::read_dir module
