+++ src/iters.rs
@@ -6,4 +6,4 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-// TODO
+pub(crate) mod read_dir;
